+++ src/delivery/Kontent.Ai.Delivery/Api/QueryBuilders/EnumerateItemsQuery.cs
@@ -176,7 +176,6 @@ internal sealed class EnumerateItemsQuery<TModel>(
                 await contentItemMapper.CompleteItemAsync(
                         item,
                         content.ModularContent,
-                        dependencyContext: null,
                         defaultRenditionPreset,
                         customAssetDomain,
                         cancellationToken)